+++ Sources/WalletConnect/Sequences/Sequence.swift
@@ -1,0 +1,7 @@
+
+import Foundation
+
+protocol Sequence: AnyObject {
+    var topic: String {get set}
+    var sequenceState: SequenceState {get set}
+}
