+++ Sources/WalletConnect/Sequences/Sequences.swift
@@ -1,7 +1,6 @@
 
 import Foundation
 
-
 class Sequences {
     var sequences: [Pairing] = []
